refactor(api)!: rename campus_id module and add docs
This commit introduces two main changes:
1.  The `info::campus_id` module is renamed to `info::ft_campus_id` to improve naming consistency across the API.
2.  Comprehensive documentation and usage examples have been added to several API endpoints, including `cursus_id_projects`, `groups`, and `projects`, to enhance developer experience.

BREAKING CHANGE: The `info::campus_id` module has been renamed to `info::ft_campus_id`. Usages of campus ID constants must be updated to the new path.
