+++ libft-api/src/api/campus/campus_id_locations.rs
@@ -115,7 +115,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::info::campus_id::GYEONGSAN;
+    use crate::info::ft_campus_id::GYEONGSAN;
 
     #[tokio::test]
     async fn location_with_params() {
